Moved some functions to helper_functions
